pyproject.toml
[project]
name = "session"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pyrogram>=2.0.106",
    "python-telegram-bot>=22.0",
    "telethon>=1.40.0",
    "tgcrypto>=1.2.5",
]
